courses/serializers/course_serializer.py
from rest_framework import serializers
from courses.models import Course


class CourseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Course
        fields = ['id', 'course_id', 'name', 'credit_hours', 'programs']
