requirements.txt
asgiref==3.8.1
Django==5.0.4
django-extensions==3.2.3
django-guardian==2.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
mysqlclient==2.2.4
pydotplus==2.0.2
pyparsing==3.1.2
sqlparse==0.5.0
tzdata==2024.1
PyJWT==2.8.0
